+++ src/main/java/no/ntnu/idata2304/group14/plantmonitor/logic/SensorUpdateCallback.java
@@ -1,6 +1,6 @@
 package no.ntnu.idata2304.group14.plantmonitor.logic;
 
 public interface SensorUpdateCallback {
-    public void updateMessageRecieved(int sensorID, int newValue);
+    public void updateMessageRecieved(int sensorID, double newValue);
 }
